+++ devices/sycl/sycl_device.cpp
@@ -210,6 +210,7 @@ OIDN_NAMESPACE_BEGIN
     case SYCLArch::XeHPC_NoDPAS: score = 20; break;
     case SYCLArch::Xe2LPG:       score = 11; break;
     case SYCLArch::Xe2HPG:       score = 21; break;
+    case SYCLArch::Xe3LPG:       score = 12; break;
     default:
       return -1;
     }
@@ -329,6 +330,7 @@ OIDN_NAMESPACE_BEGIN
         case SYCLArch::XeHPC_NoDPAS: std::cout << "Xe-HPC";  break;
         case SYCLArch::Xe2LPG:       std::cout << "Xe2-LPG"; break;
         case SYCLArch::Xe2HPG:       std::cout << "Xe2-HPG"; break;
+        case SYCLArch::Xe3LPG:       std::cout << "Xe3-LPG"; break;
         default:                     std::cout << "Unknown"; break;
         }
         std::cout << std::endl;
@@ -363,6 +365,7 @@ OIDN_NAMESPACE_BEGIN
     case SYCLArch::XeHPC:
     case SYCLArch::Xe2LPG:
     case SYCLArch::Xe2HPG:
+    case SYCLArch::Xe3LPG:
       weightDataType = DataType::Float16;
       weightLayout   = TensorLayout::OIhw8i16o2i;
       break;